requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
langchain==0.0.352
pypdf2==3.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
google-generativeai==0.3.2
requests
